+++ packages/database/src/header.js
@@ -1,6 +1,6 @@
 const name = 'home-gallery/database'
 const major = 1
-const minor = 0
+const minor = 1
 
 const isDatabaseTypeCompatible = type => type && type.startsWith(`${name}@${major}.`)